chore: organize query and mutation hooks
